split out initializers into setup.go (#6446)
